Add files via upload
This project is outline in the book "Python Crash Course, Second Edition" by Eric H. Matthes. 

This is my first time making a game in Python, and it was awesome! Definitely a lot of errors along the way, but it's in a working state. 

The book asks that you put this project in Fullscreen, and have users press "Q" to quit the game. The functionality to press "Q" to quit still remains, but I couldn't find a way to get a message to tell the user that they can press "Q" to quit. It seemed mean to put a player into fullscreen mode by default with no idea how to exit the game! So instead, I used some custom window size options. Players can press "Q" or use the exit button at the top right of the window to leave the game. Left and right arrow keys will allow players to move left and right, and "Space" bar is to shoot. Have fun.